requirements.txt
boto3==1.4.0
botocore==1.4.57
Flask==0.11.1
Flask_RESTful==0.3.5
tornado==4.4.1
Werkzeug==0.11.11
xlrd==0.9.4
pynamodb==1.5.3
PyYAML==3.12
ddt==1.1.1
